requirements.txt
Flask~=2.3.2
gradio
gunicorn
google-generativeai
requests~=2.32.3
protobuf~=5.28.2
WTForms~=3.0.1
flask_bootstrap